requirements.txt
ultralytics>=8.0.0 # As instructed in https://huggingface.co/Ultralytics/YOLOv8
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
streamlit>=1.28.0
boto3>=1.28.0
PyYAML>=6.0
tqdm>=4.65.0
ipykernel
gdown #Google Drive
python-box
ensure

